Attacks/Fuzzer/requirements.txt
pyserial==3.4
pyrecord==1.0.1
psutil==5.6.3
numpy==1.16
Flask==0.11.1
pygraphviz==1.5
colorama==0.4.1
cryptography==2.7
pycryptodome==3.8.2
socketio==0.1.7
ddt==1.2.1
mock==3.0.5
Flask-SocketIO==4.1.0
Flask==0.11.1
logbook==1.4.4
gevent==1.2.2
python-engineio==3.8.2.post1

pycallgraph==1.0.1
pygmo==2.10
socketIO-client==0.7.2
matplotlib==2.1.1